+++ lib/dom/ArrayToDOMDocument.php
@@ -66,13 +66,15 @@ class Array2DOM
 	{
 		if (!is_array($source))
 		{
+			$element = $document->createElement($tagName);
+
 			// Handle NULL bytes
 			if (strpos($source, "\0") !== false)
 			{
 				$source = 'json_encoded::' . json_encode($source);
+				$element->setAttribute('encoded', 'json');
 			}
 
-			$element = $document->createElement($tagName);
 			$element->appendChild($document->createCDATASection($source));
 
 			return $element;